+++ src/main/java/domain/card/Denomination.java
@@ -1,4 +1,4 @@
-package domain.type;
+package domain.card;
 
 public enum Denomination {
     ACE(11, "A"),
