+++ src/main/java/domain/card/Suit.java
@@ -1,4 +1,4 @@
-package domain.type;
+package domain.card;
 
 public enum Suit {
     SPADE("스페이드"), HEART("하트"), DIAMOND("다이아몬드"), CLOVER("클로버");
