+++ src/test/java/domain/PlayerTest.java
@@ -1,9 +1,11 @@
-package domain.model;
+package domain;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import domain.type.Denomination;
-import domain.type.Suit;
+import domain.card.Card;
+import domain.card.Cards;
+import domain.card.Denomination;
+import domain.card.Suit;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
